Target 'spigot-api'; NOT 'spigot'
There's nothing here that depends upon the Spigot server's non-API internals, so why pull it in?

Switching to the spigot-api pulls in a smaller dependency (3.1 MiB vs 73 MiB; 95.75% smaller!) Also, the spigot jar is no longer the server itself, but a bootstrapper containing necessary libraries. This can cause issues when developing with an Eclipse-based IDE. [See the 3rd Question in the BuildTools FAQ](https://www.spigotmc.org/wiki/buildtools/#frequently-asked-questions)

There's a maven repository for a reason.

Also, bumping the API to 1.21.1